Move function to glue_utils
